Minor improvements, readded theme stuff
git-svn-id: https://qooxdoo.svn.sourceforge.net/svnroot/qooxdoo/trunk/qooxdoo@5668 61aaf1ca-a10e-0410-b9fb-ac63ec25887f

+++ frontend/framework/tool/icon/scan-freedesktop-theme.sh
@@ -6,25 +6,30 @@ BASE=`dirname $0`
 THEMEBASE=$BASE/data/theme_
 WHITELIST="16x16 22x22 24x24 32x32 48x48 64x64 128x128 scaleable"
 CATS="actions animations apps categories devices emblems emotes intl mimetypes places status"
-DNAME=`basename $1`
 
-cd $1
-echo "  - $DNAME"
-rm -f ${THEMEBASE}${DNAME}.dat
+for DIR in $@; do
+  DNAME=`basename $DIR`
 
-for SUB in $WHITELIST; do
-  echo "    + $SUB"
+  echo "  - $DNAME"
+  rm -f ${THEMEBASE}${DNAME}.dat
 
-  for CAT in $CATS; do
-    if [ -r $DIR/$SUB/$CAT ]; then
-      for FILE in `find $DIR/$SUB/$CAT -mindepth 1 -maxdepth 1 -name "*.png" -o -name "*.svg"`; do
-        FNAME=`basename $FILE | cut -d"." -f1`
-        echo "$CAT/$FNAME" >> ${THEMEBASE}${DNAME}.dat
-      done
-    fi
+  for SUB in $WHITELIST; do
+    echo "    + $SUB"
+
+    for CAT in $CATS
+    do
+      if [ -r $DIR/$SUB/$CAT ]; then
+        for FILE in `find $DIR/$SUB/$CAT -mindepth 1 -maxdepth 1 -name "*.png" -o -name "*.svg"`; do
+          FNAME=`basename $FILE | cut -d"." -f1`
+          echo "$CAT/$FNAME" >> ${THEMEBASE}${DNAME}.dat
+        done
+      fi
+    done
   done
-done
 
-cat ${THEMEBASE}${DNAME}.dat | sort | uniq > ${THEMEBASE}${DNAME}.tmp
-rm ${THEMEBASE}${DNAME}.dat
-mv ${THEMEBASE}${DNAME}.tmp data/${THEMEBASE}${DNAME}.dat
+  if [ -r ${THEMEBASE}${DNAME}.dat ]; then
+    cat ${THEMEBASE}${DNAME}.dat | sort | uniq > ${THEMEBASE}${DNAME}.tmp
+    rm -f ${THEMEBASE}${DNAME}.dat
+    mv ${THEMEBASE}${DNAME}.tmp ${THEMEBASE}${DNAME}.dat
+  fi
+done
